Setup cucumber. Various refactors to improve test error logs

+++ api/src/employees/employees.service.ts
@@ -1,11 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import {
+  CreateEmployeeDto,
   Employee,
   EmployeeDto,
   PaginatedList,
   PaginatedListParameters,
-  Skill,
   SkilledEmployee,
+  UpdateEmployeeDto,
 } from '@skills-matrix/types';
 import { ResultSetHeader } from 'mysql2';
 import { DatabaseService } from '../database/database.service';
@@ -27,36 +28,49 @@ export class EmployeesService {
     protected readonly employeeSkillRelationsService: EmployeeSkillRelationsService,
   ) {}
 
-  async create(employeeDto: EmployeeDto): Promise<EmployeeDto> {
-    const result: ResultSetHeader = await this.databaseService.execute(insertEmployeeSql, {
-      name: employeeDto.Name,
-      surname: employeeDto.Surname,
-    });
-
-    await this.insertRelations(result.insertId, employeeDto.Skills);
-
-    return this.findOne(result.insertId);
-  }
-
-  async findAll({
-    keywords,
-    page,
-    pageSize,
-  }: PaginatedListParameters): Promise<PaginatedList<Employee>> {
+  async count(keywords?: string): Promise<number> {
     const currentKeywords = `%${keywords || ''}%`;
-    const currentPage = Number(page) || 0;
-    const currentPageSize = Number(pageSize) || 10;
-    const employeesPage: Employee[] = await this.databaseService.execute(getManyEmployeesSql, {
-      name: currentKeywords,
-      limit: String(currentPageSize),
-      offset: String(currentPage * currentPageSize),
-    });
     const [{ Total }]: [{ Total: number }] = await this.databaseService.execute(
+      'countAllEmployeesSql',
       countAllEmployeesSql,
       {
         name: currentKeywords,
       },
     );
+    return Total;
+  }
+
+  async create(employeeDto: CreateEmployeeDto): Promise<EmployeeDto> {
+    const result: ResultSetHeader = await this.databaseService.execute(
+      'insertEmployeeSql',
+      insertEmployeeSql,
+      {
+        name: employeeDto.Name,
+        surname: employeeDto.Surname,
+      },
+    );
+
+    await this.insertRelations(result.insertId, employeeDto.SkillIds);
+
+    return this.findOne(result.insertId);
+  }
+
+  async findAll({ keywords, page, pageSize }: PaginatedListParameters = {}): Promise<
+    PaginatedList<Employee>
+  > {
+    const currentKeywords = `%${keywords || ''}%`;
+    const currentPage = Number(page) || 0;
+    const currentPageSize = Number(pageSize) || 10;
+    const employeesPage: Employee[] = await this.databaseService.execute(
+      'getManyEmployeesSql',
+      getManyEmployeesSql,
+      {
+        name: currentKeywords,
+        limit: String(currentPageSize),
+        offset: String(currentPage * currentPageSize),
+      },
+    );
+    const Total = await this.count(keywords);
 
     return {
       CurrentPage: currentPage,
@@ -67,7 +81,15 @@ export class EmployeesService {
   }
 
   async findOne(id: number): Promise<EmployeeDto> {
-    const [employee] = await this.databaseService.execute(getEmployeeByIdSql, { id: String(id) });
+    const [employee] = await this.databaseService.execute(
+      'getEmployeeByIdSql',
+      getEmployeeByIdSql,
+      { id: String(id) },
+    );
+
+    if (!employee) {
+      return undefined;
+    }
 
     const skills = await this.employeeSkillRelationsService.findSkillsByEmployee(id);
 
@@ -75,30 +97,40 @@ export class EmployeesService {
   }
 
   async getMostSkilled(): Promise<SkilledEmployee[]> {
-    return this.databaseService.execute(getMostSkilledEmployeesSql);
+    return this.databaseService.execute('getMostSkilledEmployeesSql', getMostSkilledEmployeesSql);
   }
 
-  protected async insertRelations(employeeId: number, skills: Skill[]): Promise<void> {
-    for (const skill of skills) {
-      await this.employeeSkillRelationsService.create(employeeId, skill.Id);
+  protected async insertRelations(employeeId: number, skillIds: number[]): Promise<void> {
+    for (const skillId of skillIds) {
+      await this.employeeSkillRelationsService.create(employeeId, skillId);
     }
   }
 
-  async update(employeeDto: EmployeeDto): Promise<EmployeeDto> {
-    await this.databaseService.execute(updateEmployeeByIdSql, {
-      id: String(employeeDto.Id),
-      name: employeeDto.Name,
-      surname: employeeDto.Surname,
-    });
+  async update(employeeDto: UpdateEmployeeDto): Promise<EmployeeDto> {
+    const result: ResultSetHeader = await this.databaseService.execute(
+      'updateEmployeeByIdSql',
+      updateEmployeeByIdSql,
+      {
+        id: String(employeeDto.Id),
+        name: employeeDto.Name,
+        surname: employeeDto.Surname,
+      },
+    );
+
+    if (!result.affectedRows) {
+      return undefined;
+    }
 
     await this.employeeSkillRelationsService.removeByEmployeeId(employeeDto.Id);
 
-    await this.insertRelations(employeeDto.Id, employeeDto.Skills);
+    await this.insertRelations(employeeDto.Id, employeeDto.SkillIds);
 
     return this.findOne(employeeDto.Id);
   }
 
-  async remove(id: number): Promise<void> {
-    await this.databaseService.execute(deleteEmployeeByIdSql, { id: String(id) });
+  async delete(id: number): Promise<void> {
+    await this.databaseService.execute('deleteEmployeeByIdSql', deleteEmployeeByIdSql, {
+      id: String(id),
+    });
   }
 }
